sorrentum_sandbox/spring2024/SorrTask949_Basic_CRUD_Operations_with_CouchDB_RethinkDB_and_MariaDB/MariaDB/requirements.txt
Flask==2.0.3
werkzeug==2.0.3
mysql-connector-python==8.0.23
